LPS-193648 refactor CSDiagramSettingImageUploadFileEntryHandler.java and CSDiagramSettingImageUploadResponseHandler.java
move and create inner class in UploadCSDiagramSettingImageMVCActionCommand.java and delete handler class
